:truck: [#4608] Split 'default' prefill fields to separate files
Once we start re-using bits and pieces and have plugin-specific code/
configuration, it's better to stick to the one-component-per-file
approach since (some of) the components are now re-usable and semi-
public API.
